+++ src/test/java/lotto/model/PurchasePriceTest.java
@@ -1,7 +1,6 @@
 package lotto.model;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import lotto.model.domain.PurchasePrice;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
